Double checking the compression issue in C++.
The compression-disable was set differently in (win/cocos) projects, which was part of the confusion.
It should be disabled everywhere now.
It doesn't fix the Cocos Android build (it was already disabled).
Fixes CocosSDK (memcpy and such) weren't ported to WinSDK.
Definitely a huge TODO to stop having two versions of HttpRequesterCURL - only 2 functional lines different from one another.