Made multiple changes. See the full description.
Split the SDK into three different SDKs to match the others.
PlayFabClientSDK, PlayFabSDK, and PlayFabServerSDK. Make sure to use
only one version of the plugin. Use the Client plugin for the client
build and server for the server. If you build both server and client at
the same time use the full SDK.

PlayFab settings can now be set through blueprint without the need to
get into the code. Use a "Set Play Fab Settings" node. You also set the
SessionTicket through blueprint also. Set it once after authorization
using "Set Session Ticket"

In each ExampleProject there are test functions inside the level
blueprint. Just build the project, open it in the editor and navigate to
the level blueprints of the map that is loaded when opened.