Small changes with updates to the utilities functions.
The datatypes with collections as maps was being interpretted as string
arrays. Fixed it so it would make them json objects ediitable in the
engine.

Added the ability to get the error name provided the error code.  Also
added the ability to retrieve the photon application id from settings if
needed for any photon authentications.