Several changes made see description.
- Logging in with any means now automatically saves the SessionTicket to
a private variable.

- USTRUCTS for all the models have been created. Pass a request USTRUCT
into the api call. The response will be an Error USTRUCT and a response
Json. If no errors pass that Json into a decodeResponse function that
matches the api call to have a response USTRUCT created.

There might be a better way to do that but right now it works.