Made a few changes to request ustruct values, code cosmetics and more. See description.
- The ResponseData now contains the entire response Json as it did in
previous builds. The decoder portion will take that Json response and
create a USTRUCT that can you can then break in Unreal to get the
individual values.

- I converted the string array inputs to a single FString input. If you
pass a comma delimited string the code will parse it into an array
before sending the request. Also, a response string array will be parsed
back into a comma delimited string when decoded into a USTRUCT.

- Added a new function to the UPlayFabJsonObject that will set the
provided field as null. Now anytime an empty string, or a string with ""
is passed it is converted to a null value when sent as a request.

- Made some cosmetic changes to the auto generated code to make it a
little more legible to your end users. I put in comments on how things
work and what to pass into the calls. Most of it is auto generated
comments from your api.json files.