Fixing an issue with C++ model variables having the same name as the type.
Summary:
There is an issue with some languages where the variable name cannot be the same as the type-name.  C++ has an elaborate workaround to resolve this problem, but an issue was recently found with that workaround in the the model-deconstructor.
This corrects the issue so that the deconstructor deletes the variable-name, rather than the type-name.

(ZStone: This only fixes the existing C++ workaround.  It does not fix the same issue in ActionScript, so I still need you to rename the type, or the variable)

Test Plan: Existing tests caught the issue, this fixes it.

Reviewers: zstone

Reviewed By: zstone

Maniphest Tasks: T574

Differential Revision: https://phab.playfabdev.com/D241