Tests across all sdks to ensure AttributeInstall sequence submits AdvertisingId properly.
Summary:
T560
Still fixing Resharper noise, but less now
Fixing typos and minor issues for almost every sdk, for the AdvertisingId auto-sequence.
Automated testing for almost every sdk, which verifies that the AdvertisingId auto-sequence continues to work.
When possible in each sdk, the timing for the login-result is linked with the timing of the AdvertisingId result.

Test Plan: These are the tests!

Maniphest Tasks: T560

Differential Revision: https://phab.playfabdev.com/D295