SdkGenerator can utilize or ignore api calls marked with the beta flag, switching on the "-beta" command line option
Summary:
SdkGenerator can utilize or ignore api calls marked with the beta flag, switching on the "-beta" command line option
Refactoring the command line input parsing system to be cleaner and more flexible.
Removing a bunch of old tesing code that was replaced by the new Jenkins build system.
Refactoring a lot of code to be more compact and/or cleaner.
Fixing tons of resharper noise.

Test Plan: Existing tests should be sufficient

Reviewers: bvanous, #dev_tools_team, Marco.Williams

Reviewed By: #dev_tools_team, Marco.Williams

Differential Revision: https://phab.playfabdev.com/D407