Added a new C# target called 'flat'.
The goal is to minimize the amount of lines written by flattening some
of the layers involved with object creation to handle requests:
* Async requests now have both the original Request object
implementation for each method and now a second method that takes the
members of the Request object and internally constructs the Request
object on your behalf.
* These parameters are ordered from non-optional to optional with
defaults provided
* These methods have their parameters commented
* The PlayFabSettings is no longer static and can be shared amongst *API
instances. The original values are stored in PlayFabDefaultSettings and
can be modified to affect all PlayFabSettings created after changing the
defaults.
* Added a MessageHandler for hooking operations in the HttpClient object
to report back useful information such as progress.
* Explicitly added the method UploadBuildAsync to upload a build to AWS
using the url provided by GetServerBuildUploadUrl.
* Changed error handling to thow exceptions instead of returning
errors--removed the error callback.
* Unwrap results if the result container has zero or one entry in it.
* Unwrapped results with one entry also include the comment for the
entry as the return comment.
* Remove request container if it doesn't contain entries in it.
* Updated unit tests.