Removing the model inheritance from the sdks.
Summary:
Removing the model inheritance from the sdks.
Model inheritance was finished with several sdks, but now we're undoing that feature.
So, this minimally removes it in the finished sdks.
It should be easy to pick up and continue later, but for now it's cleanly removed.

Test Plan: Existing tests.

Reviewers: #devtools_in-progress, zac-playfab

Reviewed By: #devtools_in-progress, zac-playfab

Subscribers: zac-playfab

Maniphest Tasks: T1038

Differential Revision: https://phab.playfabdev.com/D1125