Fix T2117: GetPlayerStatistics API now works on UnrealBlueprints
Summary:
Fix T2117: GetPlayerStatistics API now works on UnrealBlueprints
Specifically when various TArray objects are empty, the json sent to the server is "null" instead of "[]"
Other: The ejs-templates had a bunch of logic that is traditionally stored in the make.js files in other targets.  I couldn't easily modify that logic in the ejs files, so I moved it to be more consistent and maintainable.  (I couldn't do this change without moving this logic)
fixes T2117

Test Plan: Some Jenker!, but mostly manual.

Reviewers: #devtools_in-progress, Marco.Williams

Reviewed By: #devtools_in-progress, Marco.Williams

Maniphest Tasks: T2117

Differential Revision: https://phab.playfabdev.com/D2310