Minor JS updates.
Summary:
fixes T2211
But not really.  We've decided that the JS testing is broken beyond repair, and the cost/benefit has exceeded usefulness.
The example uses the CDN which is slightly better practice than adding the script files,
and we report slightly better info about when tests finish in the example.

Test Plan: none, because we removed it :(

Reviewers: #devtools_in-progress, zac-playfab

Reviewed By: #devtools_in-progress, zac-playfab

Subscribers: zac-playfab

Maniphest Tasks: T2211

Differential Revision: https://phab.playfabdev.com/D2460