Revise the XML conversion tool to filter out messages if the test says it has otherwise passed
Summary:
The official spec for JUnit xml is that a test that posts a message is considered failed.
This is not true for most of our testing frameworks, and in the case of Unity, we're posting a passed test with a message.
This filters that message and reports the test as passed.

Also a crapton of Resharper cleaning.

Test Plan: Jenker!

Reviewers: #devtools_in-progress, Marco.Williams

Reviewed By: #devtools_in-progress, Marco.Williams

Maniphest Tasks: T2306

Differential Revision: https://phab.playfabdev.com/D2560