Fix for a PlayFab crash when the game was run on a PC with no Internet connection
Changed mUserSessionTicket back to a concrete type instead of a pointer. It was previously converted to a pointer because Aws::String had memory manager problems with shutting down statics. However now that it has been converted to an AZStd::string, it is better as a concrete type, and it fixes a crash also.

(CL 15679)