Unreal CPP is now compatible with Entity APIs and all new API methods and datatypes are working again.
Enums are restricted to 255 size max, or they fall back to being a string.
Entity and Group APIs are present.
DevSecretKey is present in all api combinations (due to entities)
Generator helper function parameter order is reset to highest-to-lowest for clarity. (I misinterpreted the meaning of the property being listed before the datatype, and subsequently expected the wrong information from datatype)
Fixed errors from the last checkin, which couldn't be fully tested.