Package testTitleData.json in UE4 for easy inclusion in mobile builds
If environment variable PF_TEST_TITLE_DATA_JSON is defined, copy it to
a Content/JSON within every UE4 ExampleProject.  Configure UE4 to
include Content/JSON in asset packaging.  This allows access to the
file from a mobile build, without requiring the developer to manually
copy testTitleData.json onto the mobile device, which is particularly
difficult on iOS.  This is also potentially useful on desktop, but
moreso on target platforms where apps are tightly sandboxed.