Disable shared libraries for debug builds
This disables a Xamarin feature that speeds up development iteration
time by deploying a couple of independent runtime APKs from Visual
Studio under the Debug configuration, instead of combining them with the
app's unique code.  The advantage to disabling this feature is that it
eliminates the possibility of a few development traps, and therefore
reduces the amount of Xamarin know-how that PlayFab developers are
required to have in order to use this test runner flexibly.