Milestone 20: Xamarin project that runs PlayFab C# unit tests and displays the results (#409)
The csharp SDK Generator target now contains XamarinTestRunner, a Visual
Studio solution for a Xamarin port of the UnittestRunner project (a
win32 console application). The purpose of this new solution is to test
the C# SDK on Android, iOS, and UWP.

A few minor changes were made to the test framework in order to make it
run properly on iOS.

The solution folder contains cmd files that produce application packages
for the various target platforms using MSBuild, and a readme that
explains the setup process required before the project can be built.
Those steps are not repeated in this documentation, please reference the
readme file. The build scripts are a supplemental alternative to
building from Visual Studio, which is also viable.

The application itself is simple, consisting of only two elements: A
button to run the tests, and a scrolling text output that displays the
test results. If an exception occurs, the exception and its stack-trace
will be displayed in the text output.

As compared to the default Xamarin template, there is a small difference
worth noting: I've disabled "use shared library", which speeds up
development iteration time by deploying a couple of independent runtime
APKs from Visual Studio under the Debug configuration, instead of
combining them with the app's unique code.  The advantage to disabling
this feature is that it eliminates the possibility of a few development
traps, and therefore reduces the amount of Xamarin know-how that PlayFab
developers are required to have in order to use this test runner
flexibly.