Moves try-catch into while loops (#423)
* Moves try-catch into while loops

This ensures an exception won't kill our process. The loop should fail on that one instance, and still pick up the next pending item.

* Catches now by const ref

* adding catch all cases

we mainly want to swallow exceptions. Display them to a user is the best case scenario, but we may not be able to in this case, so we should at least keep the background worker thread pumping.

Ideally in the future we should consider logging what happened.