Update XBO TestApp to use shared test framework
The Xbox TestApp project has been revamped to use
the new shared test framework. It will currently run through
all of the tests, perform the cloud upload, report the results
via cloud script, and return the result at the end of main.

The testTitleData.json file is copied from an environment variable
to the XBO staging area as part of the build process.

The test framework has been lightly refactored to better
support multiple platforms. TestApp is now a class, which is
instantiated and invoked from the platform-specific entry point
to the program. TestApp has also been given responsibility for
loading the test title data, and for logging output to the terminal
in a platform-dependent way.