bypass cloudscript listener in jenkins
Jenkins machines have been held up due to cloudscript limitations in the past. This commit will just output an xml file to be picked up by the cloudscript listener instead of attempting to route around an ExecuteCloudscript.

This is essentially a temporary solution. The correct solution (which will take time) would be to a.) chunk tests into separate files so ExecuteCloudscript doesn't error out with a size limitation of 10KB (but how would we get the JenkinsUtitlity to read these new chunks?), b.) upgrade our test app to allow this kind of limit (but would we just see this issue later on?).

When we start adding more tests, this problem will become much worse. This current workaround will incurr a test run cost (it takes time to output a "large" 10KB xml file. We will be adding more tests, so this file will take longer to output.