Update unity_RunAutoTests.bat (#504)
* Update unity_RunAutoTests.bat

Quick way to start packaging for PS4.

* Update unity_RunAutoTests.bat

adding a comment

* Update unity_RunAutoTests.bat

weird necessity for doing a Unity build for ps4 (without an equals sign, the project path doesn't get picked up, but this doesn't seem to be an issue for other parts of this script?)

* Update unity_RunAutoTests.bat

point ps4 to the first _BUP built project directory

* Update unity_RunAutoTests.bat

change to the correct directory before calling ps4

* Update unity_RunAutoTests.bat

pointing to _TC directory instead of _BUP

* Update unity_RunAutoTests.bat

Turn ps4 build off by default. The main build machines don't have the PS4 plugins for unity, so we need to only enable these tests on machines that do have that add on.