Convert from HttpRequestMessage to HttpRequest
This PR converts the various helpers in the CloudScript PlugIn that took
and System.Net.Http.HttpRequestMessage to use
Microsoft.AspNetCore.Http.HttpRequest, which is the recommended type for
Azure Function running on the 2.0 runtime.

Details

Add PackageReference to Microsoft.AspNetCore.Http.

Amend the various Create methods in PlayFabFunctionContexts.cs to take
an HttpRequest. Add a helper, ReadBodyAsStringAsync to do the work
previously done by HttpContent.ReadAsStringAsync.