Don't run every `if` during enum conversion.
Tiny change that _should_ be safe, but I haven't tested it.  This would
prevent enum conversion functions from running the if-check for every
possible value during conversion.  There's no code in the conversion
functions after the if-checks, so there's no reason not to leave when
a matching value is found.  See PlayFabServerDataModels.h for country
codes, currencies, generic error codes, etc. for examples of giant
functions running every if-check.

Better yet, make a bigger change to use a switch statement when
converting to json enums.  Result should be a condition-free conversion
that uses a jump table underneath for small enough enums, and falling
back on the behavior as currently written on larger enums.  Depending on
the compiler's "decision".