Update PlayFab.h.ejs
A customer reported that RF_Standalone only works in editor. It was recommended to switch to a StrongObjectPtr, but that would break our API's initial design.

Main check we need to ensure is a shared AuthenticationContext pointer across two instanced api's should "work" and not crash when the tests are done.