Toddbell/unreal ptr (#633)
* Update PlayFab.h.ejs

A customer reported that RF_Standalone only works in editor. It was recommended to switch to a StrongObjectPtr, but that would break our API's initial design.

Main check we need to ensure is a shared AuthenticationContext pointer across two instanced api's should "work" and not crash when the tests are done.

* Update PlayFab.h.ejs

narrowing if statement to only the lines that change (function should now theoretically have the same address in editor and in a cooked build).