Update PlayFabTestContext.h.ejs
LogPlayFabCommon doesn't link properly on a mac. LogTemp is commonly used throughout the Unreal codebase so this should still log the end test output, it will just label it differently as documented in the unreal wiki: https://wiki.unrealengine.com/Logs,_Printing_Messages_To_Yourself_During_Runtime

The LogPlayFabCommon is not available in the UPlayFabTestContext.h file (this file is stored in the ExampleProject/Plugin/PlayFab/Source and not under PlayFab/Source because we don't want to ship these tests with the plugin that gets distributed through the UnrealMarketplace).

We can move the declaration to the test context, but that will lose the PlayFabCommon definition. This is the quickest fix (and the wiki seems to only show using UE_LOG in a .cpp file, using it in a .h file for customer definitions may not be doable since it doesn't link on the Mac).