Toddbell/android unreal test (#652)
* Update PlayFabBlueprintTests.h.ejs

This one particular test is failing. taking it out to test if that is explicitly why no tests results are coming in the xml file.

* Update RunTestsCommandlet.cpp.ejs

GISRequestingExit is considered Deprecated as of 4.24 and must be removed by 4.25

* Update RunTestsCommandlet.cpp.ejs

adding placeholder functions

* Update DefaultEngine.ini

There's 2 DefaultEngine.ini's?!

* Create BuildAndroidOnLinux.sh.ejs

adding script for building Android on a Unix based OS

* Update RunTestsCommandlet.cpp.ejs

removing unnecessary comments

* Delete BuildAndroidOnLinux.sh.ejs

This script isn't ready yet, so it should be submitted at a later time.