Update PlayFabTestContext.h.ejs (#653)
* Update PlayFabTestContext.h.ejs

LogPlayFabCommon doesn't link properly on a mac. LogTemp is commonly used throughout the Unreal codebase so this should still log the end test output, it will just label it differently as documented in the unreal wiki: https://wiki.unrealengine.com/Logs,_Printing_Messages_To_Yourself_During_Runtime

The LogPlayFabCommon is not available in the UPlayFabTestContext.h file (this file is stored in the ExampleProject/Plugin/PlayFab/Source and not under PlayFab/Source because we don't want to ship these tests with the plugin that gets distributed through the UnrealMarketplace).

We can move the declaration to the test context, but that will lose the PlayFabCommon definition. This is the quickest fix (and the wiki seems to only show using UE_LOG in a .cpp file, using it in a .h file for customer definitions may not be doable since it doesn't link on the Mac).

* Splitting EndTest to the cpp file

If we are to explicitly label our Logs, we can only use UE_LOG in the .cpp files for cross platform development.

* Update SDKGenerator.njsproj

add the new file to the generating project

* Checking build identifier casing

It seems the generator may toLowerCase the buildIdentifier, which may be why we can't find JBuild_

* Update DefaultEngine.ini

Unreal provision needs to be updated from PlayFab to Microsoft

* Update generate.ts

removing comment

* Reverting toLower issue

the ToLowerCase build Identifier shouldn't matter. All other tests seem to do fine with it. Also adding default build settings to the EditorTarget.cs file for Unreal to get rid of noisy "Upgrade" build code. Reverting a bundle identifier that may be used by some other projects. It may need to be changed for iOS testing later.

* Update SDKGenerator.njsproj

That change shouldn't be here.

* Update generate.ts

removing excess space

* Update ACloudScriptTestResultUploader.cpp.ejs

The explicit folders seem to be required for some test includes after a BuildSettingsV2 change.

* Update ExampleProjectEditor.Target.cs

Unreal's suggested fix surprisingly broke the build! reverting their logic

* Adding mac specific build scripts

Mac may still be able to build and run tests for macOS even if iOS isn't currently working.

* Update BuildMacOS.sh.ejs

This is the accepted target platform for unreal

* Update BuildMacOS.sh.ejs

This looks promising, it popped up in command RunUAT.sh -list

* Update ACloudScriptTestResultUploader.cpp.ejs

jcu seems to auto lower case in our android test runs? But why would that be different between Windows vs Android Unreal test runs?

* Removing broken scripts

These scripts run into a dSYM issue, we may as well modify the current windows build scripts to handle the iOS push to a mac device to build/test