Toddbell/unity mac script updates (#664)
* Updating mac build scripts for unity

we made some directory changes that need to be reflected in the scripts

* Update mac_unity_RunAutoTests.sh

attempting to call Unity directly by its mac default install path

* Update mac_unity_RunAutoTests.sh

unity is stored as a .app file on a mac's Applications folder

* Update mac_unity_RunAutoTests.sh

Unity.app acts as a directory on macOS and not an exe. The actual file to run Unity is in a deeper folder.

* small notes and quickfixes

we are hard coding unity versions for the moment, also double checking Asset folder path for test scenes (may need to add the ExampleProject\ prefix)

* Update mac_unity_RunAutoTests.sh

explicitly define a scene for unity to pick up

* Update PlayFabPackager.cs

testing out universal build target (which seems to be the newer/supported build target?)

* Update mac_unity_RunAutoTests.sh

path shouldn't be required like other platforms.

* Update PlayFabPackager.cs

reverting something that didn't seem to affect the build

* Update mac_unity_RunAutoTests.sh

mac didn't explicitly call out a projectPath?

* Update mac_unity_SetupTestProjects.sh

Quick TC removal to reflect changes made in windows

* Update mac_unity_RunAutoTests.sh

Fixing comment

* Update mac_unity_RunAutoTests.sh

pulled a run command from the other Unity_RunAutoTests.sh file

* Update mac_unity_RunAutoTests.sh

run the exe without a cmd required (doesn't work on a mac)

* Update mac_unity_RunAutoTests.sh

stop and exit at failure

* Update mac_unity_RunAutoTests.sh

We are building the Unity project, but we never ran it? This should at least build a standalone version of the app?

* Update mac_unity_RunAutoTests.sh

Yeah building the player seems to have broken something else.

* Update mac_unity_RunAutoTests.sh

explicitly open the built .app file

* Update mac_unity_RunAutoTests.sh

Yeah that buildPlayer was required.

* Update mac_unity_RunAutoTests.sh

oh don't kill the job because we are outputting this. It may be a noisy false alarm.

* Update mac_unity_RunAutoTests.sh

REALLY make sure this line goes though

* Update mac_unity_RunAutoTests.sh

different build folder apparently

* Update mac_unity_RunAutoTests.sh

attempting to make the UnityVersion a more flexible parameter

* Update mac_unity_RunAutoTests.sh

Updating latest Unity versions on the mac node

* Update mac_unity_RunAutoTests.sh

Cleaning up crufty comments

* Delete mac_unity_SetupTestProjects.sh

This script is no longer called or needed.

* Update mac_unity_RunAutoTests.sh

adding notes about unity mac version issues. default to known install and optionally switch to legacy supported versions also known to be installed.