Fixing build setting warnings
I tried fixing the "We have changed the default settings so you should update that" but the fix suggested breaks previous versions of Unreal.

I tried checking for #if ENGINE_MINOR_VERSION > 23 directly in the C# Target.build file, but that didn't compile (C# didn't recognize it as a valid pragma and so thought it was not syntactically correct).

So the final conclusion is to check during the SDKGeneration process.