Update PlayFabHTTP.py
Python callbacks may have a successful call, but return an empty data block. So some successes should be considered passing. Otherwise, we will call a user failure, but they still have all the data they requested (so did the call really fail just because data may not be present in the callback?)