Reduce truncation in ADO calls
Call Builds-List for only one pipeline at a time, in order to ensure we get builds for each pipeline.
Call Pipelines-List in order to find every pipeline
Filter Pipelines-List by the ones that start with "\PlayFabSdk", which hard-locks the folder we consider "finished pipelines", but also matches current expectation
Tweak the task calling a bit.
Cache everything to files a little more thoroughly for testing.