Support AF debugging in Unity with C# SDK
Some game developers are using the PlayFab C# SDK in their Unity games.
This PR adds the UNITY_EDITOR and UNITY_STANDALONE constants to the #if
controls for the LocalServerApi setting. This will allow this setting
to work correctly via the settings file when using the PlayFab C# SDK
in a Unity game.